tests + docs
- python tests
- docs (I use AI to generate but checked by myself and fixed)
- fix in python APIs
- small changes

+++ connect/src/main/scala/org/apache/spark/sql/graphframes/GraphFramesConnectUtils.scala
@@ -457,7 +457,7 @@ object GraphFramesConnectUtils {
           rwNumWalksPerNode = message.getRwNumWalksPerNode(),
           rwBatchSize = message.getRwBatchSize(),
           rwNumBatches = message.getRwNumBatches(),
-          rwSeed = message.getRwSeed(),
+          rwSeed = message.getRwSeed().toInt,
           rwRestartProbability = message.getRwRestartProbability(),
           rwTemporaryPrefix = message.getRwTemporaryPrefix(),
           rwCachedWalks = message.getRwCachedWalks(),
@@ -477,11 +477,11 @@ object GraphFramesConnectUtils {
           word2vecNumPartitions = message.getWord2VecNumPartitions(),
           word2vecMinCount = message.getWord2VecMinCount(),
           word2vecMaxSentenceLength = message.getWord2VecMaxSentenceLength(),
-          word2vecSeed = message.getWord2VecSeed(),
+          word2vecSeed = message.getWord2VecSeed().toInt,
           word2vecStepSize = message.getWord2VecStepSize(),
           aggregateNeighbors = message.getAggregateNeighbors(),
           aggregateNeighborsMaxNbrs = message.getAggregateNeighborsMaxNbrs(),
-          aggregateNeighborsSeed = message.getAggregateNeighborsSeed())
+          aggregateNeighborsSeed = message.getAggregateNeighborsSeed().toInt)
       }
       case _ => throw new GraphFramesUnreachableException() // Unreachable
     }
